Refactor glass shader
