+++ app/src/main/res/values/strings.xml
@@ -289,7 +289,7 @@
     <!-- Shader test -->
     <string name="shader_test_title" translatable="false">Shaders test</string>
     <string name="shader_test_panel_size">Glass scale: %d\%%</string>
-    <string name="shader_test_glass_aberration">Aberration index: %.1f</string>
+    <string name="shader_test_glass_aberration">Aberration index: %.2f</string>
     <string name="shader_test_glass_refraction">Refraction index: %.2f</string>
     <string name="shader_test_glass_type">Glass type</string>
     <string name="shader_test_glass_type_sin">Curved</string>
